Polishing
This change updates a couple of types, from Byte to Integer which is strictly
less correct, but accurate with regards to the expected JSON payload.

[#101522656][resolves #499]

+++ cloudfoundry-client/src/main/java/org/cloudfoundry/client/v2/securitygroups/_RuleEntity.java
@@ -52,14 +52,14 @@ abstract class _RuleEntity {
      */
     @JsonProperty("type")
     @Nullable
-    abstract Byte getType();
+    abstract Integer getType();
 
     /**
      * The control signal for icmp
      */
     @JsonProperty("code")
     @Nullable
-    abstract Byte getCode();
+    abstract Integer getCode();
 
     /**
      * Enables logging for the egress rule
